feat(connector): add DOUBLE floating-point scalar type (RFC 0020)
Adds a fourth scalar type, DOUBLE (IEEE-754 double precision), to
duckdb_api/v1's closed set {BOOLEAN, BIGINT, VARCHAR}, including
predicate/conditional-input equality pushdown. This is the largest
single-type addition in the project's history: every value-carrying
struct in the pipeline stores one field per scalar kind side by side
rather than a tagged union, so this touches an exhaustive-switch
surface across Connector, Relational Semantics, Remote Runtime, and
Query Experience.

Canonical encoding is fixed-precision %.17g (17 significant digits,
proven to round-trip any double bit-for-bit), independently
implemented byte-identically in four layers since CMAKE_CXX_STANDARD
is 11 and std::to_chars for floating-point is C++17-only. Equality is
a direct == on the normalized (-0.0 -> 0.0) decoded double throughout,
correcting an RFC-draft-time claim (disproven by review) that existing
typed-equality functions compare via a canonical encoded string.
strtod's ERANGE fires for both true overflow and benign underflow to a
subnormal or exact zero; only HUGE_VAL/-HUGE_VAL is rejected, so a
legitimate tiny measurement still decodes.

GraphQL relations have no DOUBLE-equivalent scalar kind, so a
type: DOUBLE GraphQL column is rejected at compile time with a precise
diagnostic (mirroring RFC 0018's api_key+GraphQL precedent) rather than
silently decoding as a string.

release/1.0.0/freeze.json gains new scalar_types tracking (no existing
set to extend, unlike pagination strategies or credentials), verified
against the schema's closed column-type enum in contract_freeze.py.

Reviewed by three independent adversarial perspectives (Relational
Semantics, Transport/Remote Runtime, Test Oracle); the two P1 and two
P2 findings from the Test Oracle pass are fixed in this change (exact
subnormal round-trip assertion, a Runtime-layer underflow-acceptance
fixture, a JSON-number-spelling-equivalence test, and full DOUBLE
support in the input-resolution observation test service).

+++ test/cpp/semantics/scan_plan_response_test_fixture_tests.cpp
@@ -39,7 +39,7 @@ void TestResponseCounterexamples(const std::string &canary) {
 			Require(plan.OutputColumns().front().name.empty(), "schema-name counterexample retained a name");
 			break;
 		case ResponsePlanCounterexample::UNSUPPORTED_SCHEMA_TYPE:
-			Require(plan.OutputColumns().front().logical_type == "DOUBLE",
+			Require(plan.OutputColumns().front().logical_type == "DECIMAL",
 			        "schema-type counterexample retained a supported type");
 			break;
 		case ResponsePlanCounterexample::FLIPPED_SCHEMA_NULLABILITY:
